feat: add chartView component to sb
