+++ packages/smart-board/src/components/index.ts
@@ -1,2 +1,3 @@
 export { SmartBoardToolbar, SmartBoardToolbarProps } from './defaultToolbar';
 export { SmartBoardSelector, SmartBoardSelectorProps, SmartBoardSamples } from './defaultSelector';
+export { SmartBoardChartView, SmartBoardChartViewProps } from './defaultChartView';
