+++ sites/ava-site/examples/plugins/smart-board/demo/usage.jsx
@@ -2,11 +2,9 @@
 
 import React, { useState, useEffect } from 'react';
 import ReactDOM from 'react-dom';
-import { Tag, Tooltip } from 'antd';
 import * as G2Plot from '@antv/g2plot';
-import { LockOutlined, UnlockOutlined, MonitorOutlined } from '@ant-design/icons';
 import { statistics } from '@antv/data-wizard';
-import { SmartBoard, SmartBoardToolbar, SmartBoardSelector, smartBoardConfig } from '@antv/smart-board';
+import { SmartBoard, SmartBoardToolbar, SmartBoardSelector, SmartBoardChartView } from '@antv/smart-board';
 
 function g2plotRender(container, type, data, options) {
   const containerDOM = typeof container === 'string' ? document.getElementById(container) : container;
@@ -24,7 +22,7 @@ const cars = 'https://cdn.jsdelivr.net/npm/vega-datasets@2/data/cars.json';
 
 const chartSample1 = [
   {
-    data: cars,
+    dataUrl: cars,
     subspaces: [],
     breakdowns: ['Origin'],
     measures: ['Horsepower'],
@@ -41,7 +39,7 @@ const chartSample1 = [
     chartType: 'column_chart',
   },
   {
-    data: cars,
+    dataUrl: cars,
     subspaces: [],
     breakdowns: ['Year'],
     measures: ['Acceleration'],
@@ -50,7 +48,7 @@ const chartSample1 = [
     chartType: 'line_chart',
   },
   {
-    data: cars,
+    dataUrl: cars,
     subspaces: [],
     breakdowns: ['Origin'],
     measures: ['Miles_per_Gallon'],
@@ -59,7 +57,7 @@ const chartSample1 = [
     chartType: 'pie_chart',
   },
   {
-    data: cars,
+    dataUrl: cars,
     subspaces: [],
     breakdowns: ['Cylinders', 'Origin'],
     measures: ['Displacement'],
@@ -68,7 +66,7 @@ const chartSample1 = [
     chartType: 'grouped_column_chart',
   },
   {
-    data: cars,
+    dataUrl: cars,
     subspaces: [],
     breakdowns: ['Year', 'Origin'],
     measures: ['Weight_in_lbs'],
@@ -77,7 +75,7 @@ const chartSample1 = [
     chartType: 'line_chart',
   },
   {
-    data: cars,
+    dataUrl: cars,
     subspaces: [],
     breakdowns: ['Year'],
     measures: ['Displacement'],
@@ -91,7 +89,7 @@ const gapminder = 'https://cdn.jsdelivr.net/npm/vega-datasets@2/data/gapminder.j
 
 const chartSample2 = [
   {
-    data: gapminder,
+    dataUrl: gapminder,
     subspaces: [],
     breakdowns: ['country'],
     measures: ['fertility'],
@@ -108,7 +106,7 @@ const chartSample2 = [
     chartType: 'column_chart',
   },
   {
-    data: gapminder,
+    dataUrl: gapminder,
     subspaces: [],
     breakdowns: ['year'],
     measures: ['pop'],
@@ -117,7 +115,7 @@ const chartSample2 = [
     chartType: 'line_chart',
   },
   {
-    data: gapminder,
+    dataUrl: gapminder,
     subspaces: [],
     breakdowns: ['cluster'],
     measures: ['fertility'],
@@ -126,7 +124,7 @@ const chartSample2 = [
     chartType: 'pie_chart',
   },
   {
-    data: gapminder,
+    dataUrl: gapminder,
     subspaces: [],
     breakdowns: ['country', 'cluster'],
     measures: ['pop'],
@@ -135,7 +133,7 @@ const chartSample2 = [
     chartType: 'grouped_column_chart',
   },
   {
-    data: gapminder,
+    dataUrl: gapminder,
     subspaces: [],
     breakdowns: ['year', 'country'],
     measures: ['pop'],
@@ -144,7 +142,7 @@ const chartSample2 = [
     chartType: 'line_chart',
   },
   {
-    data: gapminder,
+    dataUrl: gapminder,
     subspaces: [],
     breakdowns: ['year'],
     measures: ['pop'],
@@ -155,122 +153,6 @@ const chartSample2 = [
 ];
 
 const CHART_SAMPLE_LIST = [chartSample1, chartSample2];
-
-const ChartView = ({ chartID, chartInfo, clusterID, interactionMode, hasLocked, changeConnectionID, quitResort }) => {
-  const [curChartConfig, setChartConfig] = useState();
-  let plot;
-
-  useEffect(() => {
-    fetch(chartInfo.data)
-      .then((res) => {
-        return res.json();
-      })
-      .then((data) => {
-        const chartConfig = smartBoardConfig(chartInfo, data);
-        const { xField, yField, colorField, angleField, seriesField } = chartConfig.config;
-
-        let aggregatedData = data;
-        if ((xField || colorField) && (yField || angleField)) {
-          aggregatedData = statistics.aggregate(
-            data,
-            xField || colorField || '',
-            yField || angleField || '',
-            seriesField
-          );
-        }
-
-        setChartConfig(chartConfig);
-        plot = g2plotRender(`chart_container_${chartID}`, chartConfig.type, aggregatedData, chartConfig.config);
-      });
-    return function cleanup() {
-      if (plot) {
-        plot.destroy();
-      }
-    };
-  }, [chartID, chartInfo, hasLocked]);
-
-  const [chartClassName, setChartClassName] = useState('chart-view');
-  useEffect(() => {
-    setChartClassName(`chart_view ${interactionMode === 'clusterMode' ? clusterID : ''}`);
-  }, [interactionMode]);
-
-  const [isLocked, toggleLocked] = useState(false);
-  const handleResort = () => {
-    changeConnectionID(chartID);
-    toggleLocked(!isLocked);
-  };
-
-  const cancelResort = () => {
-    quitResort();
-    toggleLocked(!isLocked);
-  };
-
-  const lockIcon = isLocked ? (
-    <Tooltip title={'Cancel Connection'}>
-      <LockOutlined
-        className="resort_icon"
-        style={{ color: 'red', visibility: interactionMode === 'connectionMode' ? 'visible' : 'hidden' }}
-        onClick={cancelResort}
-      />
-    </Tooltip>
-  ) : (
-    <Tooltip title={'Show Connection'}>
-      <UnlockOutlined
-        className="resort_icon"
-        style={{ visibility: interactionMode === 'connectionMode' && !hasLocked ? 'visible' : 'hidden' }}
-        onClick={handleResort}
-      />
-    </Tooltip>
-  );
-
-  const config = curChartConfig?.config;
-  const dimension =
-    curChartConfig?.type !== 'Pie' ? `${config?.xField} ${config?.seriesField || ''}` : config?.colorField;
-  const measure = curChartConfig?.type !== 'Pie' ? config?.yField : config?.angleField;
-  const score = curChartConfig?.score;
-  const { description } = chartInfo;
-
-  const linkTag = {
-    SAME_DIMENSION: 'SD',
-    SAME_MEASURE: 'SM',
-    SAME_INSIGHT_TYPE: 'SI',
-  };
-
-  return (
-    <div className={chartClassName} id={`chart_view_${chartID}`}>
-      <div className="title_view">
-        <div className="title_info">
-          {score && (
-            <Tooltip title={`Score: ${score}`}>
-              <Tag icon={<MonitorOutlined />} color="error">{`${score}`}</Tag>
-            </Tooltip>
-          )}
-          {dimension && (
-            <Tooltip title={`Dimension: ${dimension}`}>
-              <Tag color="processing">{`${dimension}`}</Tag>
-            </Tooltip>
-          )}
-          {measure && (
-            <Tooltip title={`Measure: ${measure}`}>
-              <Tag color="success">{`${measure}`}</Tag>
-            </Tooltip>
-          )}
-        </div>
-        <div className="right_icons">{lockIcon}</div>
-      </div>
-      <div id={`chart_container_${chartID}`}></div>
-      {hasLocked &&
-        description &&
-        description.map((d) => {
-          return (
-            <Tooltip key={d} title={d}>
-              <Tag>{linkTag[d]}</Tag>
-            </Tooltip>
-          );
-        })}
-    </div>
-  );
-};
 
 const Dashboard = ({ chartList, interactionMode }) => {
   const smartBoard = new SmartBoard(chartList);
@@ -320,13 +202,15 @@ const Dashboard = ({ chartList, interactionMode }) => {
       {curChartList.map((chart) => {
         const clusterIndex = chartCluster[chart.id];
         return (
-          <ChartView
+          <SmartBoardChartView
             key={chart.id}
             chartID={chart.id}
             chartInfo={chart}
             interactionMode={interactionMode}
             clusterID={`cluster_${clusterIndex}`}
             hasLocked={!!connectionID} // if there exist connectionID, it means the dashboard comes into connection view
+            aggregate={statistics.aggregate}
+            g2plotRender={g2plotRender}
             changeConnectionID={changeConnectionID}
             quitResort={quitResort}
           />
